src/backend/kingfisher/backend/requirements.txt
# KingFisher Module Dependencies

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Telegram integration
telethon==1.32.1
python-telegram-bot==20.7

# Image processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Configuration
pydantic==2.5.0
python-dotenv==1.0.0

# Monitoring
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0 